Made an external library of pipeline dump for the compiler.
- Removed dependency of pipeline dump on llpc
- 1) Seperated out code for pipeline dump from llpcUtil.* to vkgcUtil.*
- 2) Moved some structs llpc.h to vkgcDefs.h and used #include "vkgcDefs.h" for dump
- 3) Removed llpcElfWriter as class friend of llpcElfReader
- 4) Added atimake and cmake files in tool/pipelineDump to create the library and modified the llpc atimake and cmake file to link the library
