+++ llpc/tool/Makefile.apicompilerllpctool
@@ -117,7 +117,8 @@ ifdef METROHASH_PATH
 endif
 
 LLLIBS += ../../$(BUILD_DIR)/icdapicompilerllpc$(LIB_EXT) \
-          ../../../tool/vfx/$(BUILD_DIR)/icdapicompilertoolvfx$(LIB_EXT)
+          ../../../tool/vfx/$(BUILD_DIR)/icdapicompilertoolvfx$(LIB_EXT) \
+          ../../../tool/pipelineDump/$(BUILD_DIR)/icdapicompilertoolpipelineDump$(LIB_EXT)
 
 ifeq ($(BUILD_LLVM_SHARED),1)
     LLOPTS += $(shell cat ../../../../../imported/llvm/$(BUILD_DIR)/link_options)
