+++ llpc/util/llpcShaderModuleHelper.cpp
@@ -32,6 +32,7 @@
 #include "llpcDebug.h"
 #include "llpcUtil.h"
 #include "spirvExt.h"
+#include "vkgcUtil.h"
 #include <unordered_set>
 #include <set>
