+++ llpc/util/llpcUtil.cpp
@@ -38,9 +38,6 @@
 
 #include "palPipelineAbi.h"
 
-#include <sys/types.h>
-#include <sys/stat.h>
-
 #define DEBUG_TYPE "llpc-util"
 
 namespace Llpc
@@ -77,57 +74,6 @@ const char* GetShaderStageName(
     }
 
     return pName;
-}
-
-// =====================================================================================================================
-// Gets name string of the abbreviation for the specified shader stage
-const char* GetShaderStageAbbreviation(
-    ShaderStage shaderStage,  // Shader stage
-    bool        upper)        // Whether to use uppercase for the abbreviation (default is lowercase)
-{
-    const char* pAbbr = nullptr;
-
-    if (shaderStage == ShaderStageCopyShader)
-    {
-        pAbbr = upper ? "COPY" : "Copy";
-    }
-    else if (shaderStage < ShaderStageCount)
-    {
-        if (upper)
-        {
-            static const char* ShaderStageAbbrs[] =
-            {
-                "VS",
-                "TCS",
-                "TES",
-                "GS",
-                "FS",
-                "CS",
-            };
-
-            pAbbr = ShaderStageAbbrs[static_cast<uint32_t>(shaderStage)];
-        }
-        else
-        {
-            static const char* ShaderStageAbbrs[] =
-            {
-                "Vs",
-                "Tcs",
-                "Tes",
-                "Gs",
-                "Fs",
-                "Cs",
-            };
-
-            pAbbr = ShaderStageAbbrs[static_cast<uint32_t>(shaderStage)];
-        }
-    }
-    else
-    {
-        pAbbr = "Bad";
-    }
-
-    return pAbbr;
 }
 
 // =====================================================================================================================
@@ -193,47 +139,6 @@ uint32_t ShaderStageToMask(
     return (1 << stage);
 }
 
-// =====================================================================================================================
-// Create directory.
-bool CreateDirectory(
-    const char* pDir)  // [in] the path of directory
-{
-    int result = mkdir(pDir, S_IRWXU);
-    return (result == 0);
-}
-
-// =====================================================================================================================
-// Helper macro
-#define CASE_CLASSENUM_TO_STRING(TYPE, ENUM) \
-    case TYPE::ENUM: pString = #ENUM; break;
-
-// =====================================================================================================================
-// Translate enum "ResourceMappingNodeType" to string
-const char* GetResourceMappingNodeTypeName(
-    ResourceMappingNodeType type)  // Resource map node type
-{
-    const char* pString = nullptr;
-    switch (type)
-    {
-    CASE_CLASSENUM_TO_STRING(ResourceMappingNodeType, Unknown)
-    CASE_CLASSENUM_TO_STRING(ResourceMappingNodeType, DescriptorResource)
-    CASE_CLASSENUM_TO_STRING(ResourceMappingNodeType, DescriptorSampler)
-    CASE_CLASSENUM_TO_STRING(ResourceMappingNodeType, DescriptorYCbCrSampler)
-    CASE_CLASSENUM_TO_STRING(ResourceMappingNodeType, DescriptorCombinedTexture)
-    CASE_CLASSENUM_TO_STRING(ResourceMappingNodeType, DescriptorTexelBuffer)
-    CASE_CLASSENUM_TO_STRING(ResourceMappingNodeType, DescriptorFmask)
-    CASE_CLASSENUM_TO_STRING(ResourceMappingNodeType, DescriptorBuffer)
-    CASE_CLASSENUM_TO_STRING(ResourceMappingNodeType, DescriptorTableVaPtr)
-    CASE_CLASSENUM_TO_STRING(ResourceMappingNodeType, IndirectUserDataVaPtr)
-    CASE_CLASSENUM_TO_STRING(ResourceMappingNodeType, PushConst)
-    CASE_CLASSENUM_TO_STRING(ResourceMappingNodeType, DescriptorBufferCompact)
-    CASE_CLASSENUM_TO_STRING(ResourceMappingNodeType, StreamOutTableVaPtr)
-        break;
-    default:
-        llvm_unreachable("Should never be called!");
-        break;
-    }
-    return pString;
 }
 
 } // Llpc
