+++ llpc/util/llpcUtil.h
@@ -38,9 +38,6 @@
 namespace Llpc
 {
 
-// Invalid value
-static const uint32_t InvalidValue  = ~0u;
-
 // Size of vec4
 static const uint32_t SizeOfVec4 = sizeof(float) * 4;
 
@@ -51,9 +48,6 @@ static const uint32_t DescSetMask           = 0x000000FF;
 
 // Gets the name string of shader stage.
 const char* GetShaderStageName(ShaderStage shaderStage);
-
-// Gets name string of the abbreviation for the specified shader stage.
-const char* GetShaderStageAbbreviation(ShaderStage shaderStage, bool upper = false);
 
 // Translates shader stage to corresponding stage mask.
 uint32_t ShaderStageToMask(ShaderStage stage);
@@ -71,24 +65,6 @@ inline uint32_t GetModuleIdByIndex(
 {
     static const uint32_t BaseModuleId = 1;
     return BaseModuleId + index;
-}
-
-// Create directory.
-bool CreateDirectory(const char* pDir);
-
-// Translate enum "ResourceMappingNodeType" to string
-const char* GetResourceMappingNodeTypeName(ResourceMappingNodeType type);
-
-// =====================================================================================================================
-// Increments a pointer by nBytes by first casting it to a uint8_t*.
-//
-// Returns incremented pointer.
-inline void* VoidPtrInc(
-    const void* p,         // [in] Pointer to be incremented.
-    size_t      numBytes)  // Number of bytes to increment the pointer by
-{
-    void* ptr = const_cast<void*>(p);
-    return (static_cast<uint8_t*>(ptr) + numBytes);
 }
 
 // =====================================================================================================================
@@ -143,35 +119,6 @@ inline uint32_t FloatToBits(
     float f)        // Float to be converted to bits
 {
     return (*(reinterpret_cast<uint32_t*>(&f)));
-}
-
-// ===================================================================================
-// Finds the expected structure in Vulkan structure chain with the specified info.
-template<class T>
-inline const T* FindVkStructInChain(
-    VkStructureType type,    // Vulkan structure type
-    const void*     pNext)   // Base pointer of Vulkan structure
-{
-    struct VkStructHeader
-    {
-        VkStructureType type;
-        VkStructHeader* pNext;
-    };
-
-    const VkStructHeader* pStructHeader = reinterpret_cast<const VkStructHeader*>(pNext);
-    while(pStructHeader != nullptr)
-    {
-        if (pStructHeader->type == type)
-        {
-            break;
-        }
-        else
-        {
-            pStructHeader = pStructHeader->pNext;
-        }
-    }
-
-    return reinterpret_cast<const T*>(pStructHeader);
 }
 
 } // Llpc
